+++ src/test/java/com/ctm/web/health/lhc/calculation/NeverHeldCoverCalculatorTest.java
@@ -11,13 +11,13 @@ public class NeverHeldCoverCalculatorTest {
 
     @Test
     public void givenAge44_thenCalculateLHC() {
-        int calculate = new NeverHeldCoverCalculator(44).calculateLHCPercentage();
+        long calculate = new NeverHeldCoverCalculator(44).calculateLHCPercentage();
         assertEquals(28, calculate);
     }
 
     @Test
     public void givenAgeBelow30_thenReturn0() {
-        int calculate = new NeverHeldCoverCalculator(25).calculateLHCPercentage();
+        long calculate = new NeverHeldCoverCalculator(25).calculateLHCPercentage();
         assertEquals(0, calculate);
     }
